refactor: split tunnelgateway impl.go by responsibility (#111)
Move HTTP/WebSocket proxy, session registry, P2P, health checks, and lifecycle into focused files; impl.go retains types and construction only.
